Luis - Clarify YANG-based service API description
Updated wording for clarity and consistency in the description of YANG-based service APIs and their operational relevance.

Luis: What I miss is maybe a reference to external systems in the sense that the purpose of those APIs (or part of them) could be for making the (IETF) network consumable. With such explicit mention we can provide evidence (mainly to people external to IETF) that ONIONS would be the channel for working on such interplay between IETF world and outside world.